# Optimize `StyleBoxFlat.draw()`