I'm not great at optimizing, but I needed this... I don't know which changes contributed most to the speedup, but resizing beforehand and the unfolding of the last nested loop seem like the primary two.

```gdscript
# Example of how I tested this
func _ready() -> void:
	var sbf := StyleBoxFlat.new()
	sbf.border_color = Color.RED
	sbf.set_corner_radius_all(4)
	var t := Time.get_ticks_msec()
	for i in 1000:
		sbf.draw(get_canvas_item(), Rect2(10, 10, 60, 20))
	print(Time.get_ticks_msec() - t)
```

**Benchmarks:**

- Rounded stylebox: 205ms on master, 110ms in my PR (1.9x faster)
- Stylebox with sharp corners: 75ms on master, 65ms in my PR (1.2x faster)

On production builds, on 4.3.beta2 I get 58ms on average, on this artifact I got 39ms on average.

This speedup should make stylebox drawing faster, for example if you draw hundreds of styleboxes in the same frame. It won't affect stylebox rendering after they have been drawn.