@AThousandShips 2024-07-12
I'd suggest using `ptrw` to improve iteration if we're resizing