@MewPurPur 2024-07-12
> I'd suggest using `ptrw` to improve iteration if we're resizing

Not too familiar with pointers, but I think I got it? I got slightly better measures, but I don't think they are statistically-significant