@MewPurPur 2024-07-12
Changing title as I discovered a notable speedup in drawing styleboxes in general.