@MewPurPur 2024-09-02
I expected that some recent optimizations, like the Vector.insert() optimization, might change the benchmark. But I still measure the same 1.9x speedup for rounded styleboxes and 1.2x for all others.